Fixing issue with path
