+++ tools/zephyr-bundle/native-packages/mac/gradle.properties
@@ -1,5 +1,5 @@
-warpDownloadURL=https://github.com/dgiagio/warp/releases/download/v0.3.0/linux-x64.warp-packer
+warpDownloadURL=https://github.com/dgiagio/warp/releases/download/v0.3.0/macos-x64.warp-packer
 wineDownloadURL=https://github.com/sunshower-io/zephyr/releases/download/minwine-windows/minwine-windows.zip
 rcEditDownloadURL=https://github.com/electron/rcedit/releases/download/v1.1.1/rcedit-x64.exe
-binPath=exe/linux
+binPath=exe/mac
 group=io.zephyr.bundle
